ast-mdio: Migrate to Python3
This change is a bit confusing since ast-mdio used to
work with python2, but the python package has been
blacklisted, and also the dependencies have changed
with the OE-Core releases.

The ast-mdio script relies on subprocess, time and
argparse to work properly, and these are all coming
from python3-core now.

Change RDEPENDS_ast-mdio to python3-core for the
reasons stated above, and also change the shebang
of the script to use python3 instead to avoid QA
errors regarding runtime dependencies while packaging.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <aehs29@gmail.com>