u-boot-tools: Add dtc-native dependency
When IMAGE_FSTYPES contains cpio the image_cpio task is executed,
this causes images to fail showing the following error:

| DEBUG: Executing shell function do_image_cpio
| 151619 blocks
| 2 blocks
| sh: dtc: command not found
| mkimage: Can't open fit-wedge-<date>.itb.tmp: No such file or
  directory
| WARNING: temp/run.do_image_cpio. exit 255 from 'mkimage -f
  recipe-sysroot/etc/fit-wedge.its ${FIT_DESTINATION}'
| ERROR: Function failed: do_image_cpio

This is caused due to a missing dependency from the native
u-boot-tools package to dtc-native, which tries to use it but
it is unavailable.

Create a bbappend to the u-boot-tools package to inject this
runtime dependency which allows mkimage to run correctly
fixing the issue.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <aehs29@gmail.com>