u-boot: Fix compilation error due to incompatibility with dtc
DTC does not use an underscore for the LIBFDT_H and
LIBFDT_ENV_H variables, this causes an error since u-boot does.

Add a patch that removes the underscore to allow compilation
to succeed.

This patch is only required on the warrior release due to the
version of dtc upgrade.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <aehs29@gmail.com>