Upgrade ipmitool to use git repo instead
The current implementation of ipmitool v1.8.18 contains
arcaic code that is incompatible with newer C compilers,
the last release was on Oct 2016 and several fixes,
(around 150 commits) have been pushed since.

Build ipmitool using the git repo instead, to avoid
compilation errors.
The FB plugin implementation for obmc using a patch
was rebased to work on the latest revision

Signed-off-by: Alejandro Enedino Hernandez Samaniego <aehs29@gmail.com>