fw-util: Port the fw-util and fw-util-test recipes to meta-fbttn
Given that the BBFILE_PRIORITY of the common layer is set to 6
and the one for meta-fbttn is set to 20, the recipe fw-util is
being forced to be taken from the meta-fbttn layer (which only
contains v0.1 and fails to build at this point).

This used to work on previous poky branches but the implementation
of BBFILE_PRIORITY changed / was fixed on newer branches, which
allows us to see this error that we were not able to see before
even if it already existed.

The following is taken from the definition of BBFILE_PRIORITY in
the Yocto reference manual:

"The precedence established through this variable stands regardless of a
recipe's version (PV variable). For example, a layer that has a recipe
with a higher PV value but for which the BBFILE_PRIORITY is set to have
a lower precedence still has a lower precedence."

Since the logical way is still to keep meta-fbttn with a higher
BBFILE_PRIORITY, we have to port the recipe from the common layer
to meta-fbttn to be able to build it correctly (v0.2)

Signed-off-by: Alejandro Enedino Hernandez Samaniego <aehs29@gmail.com>