fw-util: Port the fw-util and fw-util-test recipes to meta-fby2
Given that the BBFILE_PRIORITY of the common layer is set to 6
and the one for meta-fby2 is set to 8, the recipe fw-util is
being forced to be taken from the meta-fby2 layer (which only
contains v0.1 and fails to build at this point).

This used to work on previous poky branches but the implementation
of BBFILE_PRIORITY changed / was fixed on newer branches, which
allows us to see this error that we were not able to see before
even if it already existed.

The following is taken from the definition of BBFILE_PRIORITY in
the Yocto reference manual:

"The precedence established through this variable stands regardless of a
recipe's version (PV variable). For example, a layer that has a recipe
with a higher PV value but for which the BBFILE_PRIORITY is set to have
a lower precedence still has a lower precedence."

Since the logical way is still to keep meta-fby2 with a higher
BBFILE_PRIORITY, we have to port the recipe from the common layer
to meta-fby2 to be able to build it correctly (v0.2)

In case the v0.1 still wants to be kept, the perhaps it would still
be a better idea to create a dummy v0.2 on the common layer and
append the required contents on each of the other layers, and handle
the rest using PREFERRED_VERSION.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <aehs29@gmail.com>